change reserve to uint128

+++ contracts/core/ITermMaxMarket.sol
@@ -247,7 +247,7 @@ interface ITermMaxMarket {
     function setLsf(uint32 lsf) external;
     
     /// @notice Return the reserves of FT and XT
-    function ftXtReserves() external view returns (uint256 ftReserve, uint256 xtReserve);
+    function ftXtReserves() external view returns (uint128 ftReserve, uint128 xtReserve);
 
     /// @notice Return the tokens in TermMax Market
     /// @return ft Fixed-rate Token(bond token). Earning Fixed Income with High Certainty
